+++ snomed/com.b2international.snowowl.snomed.reasoner/src/com/b2international/snowowl/snomed/reasoner/request/OntologyExportRequest.java
@@ -85,6 +85,7 @@ final class OntologyExportRequest implements Request<BranchContext, String>, Bra
 		taxonomyBuilder.addActiveConceptIds(revisionSearcher);
 		taxonomyBuilder.finishConcepts();
 		
+		taxonomyBuilder.addFullySpecifiedNames(revisionSearcher);
 		taxonomyBuilder.addConceptFlags(revisionSearcher);
 		taxonomyBuilder.addActiveStatedEdges(revisionSearcher);
 		taxonomyBuilder.addActiveStatedNonIsARelationships(revisionSearcher);
@@ -106,7 +107,7 @@ final class OntologyExportRequest implements Request<BranchContext, String>, Bra
 			final OWLOntology ontology = ontologyManager.createOntology(ontologyIRI);
 			OWLOntology ontologyToExport = ontologyManager.createOntology();
 			
-			ontology.getLogicalAxioms().forEach(axiom -> {
+			ontology.getAxioms().forEach(axiom -> {
 				ontologyManager.addAxiom(ontologyToExport, axiom);
 			});