+++ snomed/com.b2international.snowowl.validation.snomed/src/main/resources/scripts/rule_duplicate_members.groovy
@@ -59,7 +59,9 @@ memberAggregation.getBuckets().entrySet().each({entry ->
 				// ignore
 				break;
 		}
-		issues.add(affectedComponent)
+		if (!issues.contains(affectedComponent)) {
+			issues.add(affectedComponent)
+		}
 	})
 })