+++ snomed/com.b2international.snowowl.validation.snomed/src/main/resources/scripts/rule_mrcm_constraint.groovy
@@ -1,5 +1,7 @@
 package scripts
 
+import java.util.stream.Collectors
+
 import com.b2international.index.Hits
 import com.b2international.index.query.Expression
 import com.b2international.index.query.Expressions
@@ -9,13 +11,17 @@ import com.b2international.index.revision.RevisionSearcher
 import com.b2international.snowowl.core.ComponentIdentifier
 import com.b2international.snowowl.core.date.EffectiveTimes
 import com.b2international.snowowl.snomed.common.SnomedTerminologyComponentConstants
+import com.b2international.snowowl.snomed.common.SnomedConstants.Concepts
 import com.b2international.snowowl.snomed.core.domain.constraint.SnomedCardinalityPredicate
 import com.b2international.snowowl.snomed.core.domain.constraint.SnomedConceptSetDefinition
 import com.b2international.snowowl.snomed.core.domain.constraint.SnomedConstraint
 import com.b2international.snowowl.snomed.core.domain.constraint.SnomedRelationshipPredicate
+import com.b2international.snowowl.snomed.core.domain.refset.SnomedReferenceSetMember
 import com.b2international.snowowl.snomed.core.ecl.EclExpression
 import com.b2international.snowowl.snomed.core.tree.Trees
 import com.b2international.snowowl.snomed.datastore.index.entry.SnomedConceptDocument
+import com.b2international.snowowl.snomed.datastore.index.entry.SnomedOWLRelationshipDocument
+import com.b2international.snowowl.snomed.datastore.index.entry.SnomedRefSetMemberIndexEntry
 import com.b2international.snowowl.snomed.datastore.index.entry.SnomedRelationshipIndexEntry
 import com.b2international.snowowl.snomed.datastore.request.SnomedRelationshipSearchRequestBuilder
 import com.b2international.snowowl.snomed.datastore.request.SnomedRequests
@@ -26,17 +32,22 @@ import com.google.common.collect.Multimap
 import com.google.common.collect.Sets
 
 RevisionSearcher searcher = ctx.service(RevisionSearcher.class)
-List<ComponentIdentifier> issues = Lists.newArrayList()
+Set<ComponentIdentifier> issues = Sets.newHashSet()
+
 String oneMandatoryIsAPerConceptRuleId = "b93819e3-2679-46f6-a35d-3a749842d83e"
+
+def getPredicate = { SnomedConstraint constraint ->
+	return constraint.getPredicate() instanceof SnomedCardinalityPredicate
+			? ((SnomedCardinalityPredicate) constraint.getPredicate()).getPredicate()
+			: constraint.getPredicate()
+}
 
 def mrcmRules = SnomedRequests.prepareSearchConstraint()
 		.all()
 		.build()
 		.execute(ctx)
 		.stream()
-		.filter({SnomedConstraint constraint -> constraint.getPredicate() instanceof SnomedCardinalityPredicate
-			? ((SnomedCardinalityPredicate) constraint.getPredicate()).getPredicate() instanceof SnomedRelationshipPredicate
-			: constraint.getPredicate() instanceof SnomedRelationshipPredicate})
+		.filter({SnomedConstraint constraint -> getPredicate(constraint) instanceof SnomedRelationshipPredicate})
 		.filter({SnomedConstraint constraint -> !constraint.getId().equals(oneMandatoryIsAPerConceptRuleId)})
 		.collect();
 		
@@ -63,24 +74,30 @@ def getApplicableConcepts = { String conceptSetExpression ->
 	searcher.search(conceptSetQuery)
 		.each({id -> conceptIds.add(id)})
 	return conceptIds
-}
-
-def getCachedApplicableConcepts = { String conceptSetExpression ->
-	return getApplicableConcepts(conceptSetExpression)
 }.memoize()
-		
+
 def getApplicableRules = { String conceptId ->
 	final Set<SnomedConstraint> applicableConstraints = Sets.newHashSet()
 	for (SnomedConstraint constraint: mrcmRules) {
 		
 		SnomedConceptSetDefinition domain = constraint.domain
-		if (getCachedApplicableConcepts(domain.toEcl()).contains(conceptId)) {
+		if (getApplicableConcepts(domain.toEcl()).contains(conceptId)) {
 			applicableConstraints.add(constraint)
 		}
 	}
 	
 	return applicableConstraints
 }.memoize()
+
+def getOWLRelationships = { SnomedReferenceSetMember owlMember ->
+	def owlRelationships = [] as List
+	if (owlMember.getClassOWLRelationships() != null) {
+		owlRelationships.addAll(owlMember.getClassOWLRelationships());
+	} else if (owlMember.getEquivalentOWLRelationships() != null) {
+		owlRelationships.addAll(owlMember.getEquivalentOWLRelationships());
+	}
+	return owlRelationships
+}
 
 if (params.isUnpublishedOnly) {
 	SnomedRelationshipSearchRequestBuilder  requestBuilder = SnomedRequests.prepareSearchRelationship()
@@ -91,15 +108,15 @@ if (params.isUnpublishedOnly) {
 	Expression where = requestBuilder.build().prepareQuery(ctx);
 
 	final Query<String[]> relationshipQuery = Query.select(String[].class)
-	.from(SnomedRelationshipIndexEntry.class)
-	.fields(SnomedRelationshipIndexEntry.Fields.ID,
-		SnomedRelationshipIndexEntry.Fields.SOURCE_ID,
-		SnomedRelationshipIndexEntry.Fields.DESTINATION_ID,
-		SnomedRelationshipIndexEntry.Fields.TYPE_ID,
-		SnomedRelationshipIndexEntry.Fields.CHARACTERISTIC_TYPE_ID)
-	.where(where)
-	.limit(Integer.MAX_VALUE)
-	.build();
+		.from(SnomedRelationshipIndexEntry.class)
+		.fields(SnomedRelationshipIndexEntry.Fields.ID,
+			SnomedRelationshipIndexEntry.Fields.SOURCE_ID,
+			SnomedRelationshipIndexEntry.Fields.DESTINATION_ID,
+			SnomedRelationshipIndexEntry.Fields.TYPE_ID,
+			SnomedRelationshipIndexEntry.Fields.CHARACTERISTIC_TYPE_ID)
+		.where(where)
+		.limit(Integer.MAX_VALUE)
+		.build();
 
 	Iterable<Hits<String[]>> queryResult = ctx.service(RevisionSearcher.class).search(relationshipQuery);
 
@@ -113,29 +130,51 @@ if (params.isUnpublishedOnly) {
 		def applicableRules = getApplicableRules(sourceId)
 
 		for (SnomedConstraint constraint : applicableRules) {
-			SnomedRelationshipPredicate predicate
-			if (constraint.getPredicate() instanceof SnomedCardinalityPredicate) {
-				predicate = ((SnomedCardinalityPredicate) constraint.getPredicate()).getPredicate()
-			} else {
-				predicate = constraint.getPredicate()
-			}
-
+			SnomedRelationshipPredicate predicate = getPredicate(constraint)
+			
 			def predicateCharType = predicate.getCharacteristicTypeId()
 			if (predicate.getAttributeExpression().equals(typeId) && (Strings.isNullOrEmpty(predicateCharType) || charTypeId.equals(predicateCharType))) {
-				if (!getCachedApplicableConcepts(predicate.getRangeExpression()).contains(destinationId)) {
+				if (!getApplicableConcepts(predicate.getRangeExpression()).contains(destinationId)) {
 					issues.add(ComponentIdentifier.of(SnomedTerminologyComponentConstants.RELATIONSHIP_NUMBER, relationshipId))
 				}
 			}
 		}
 	}
+	
+	def owlAxiomMembers = SnomedRequests.prepareSearchMember()
+	 	.filterByRefSet(Concepts.REFSET_OWL_AXIOM)
+	 	.filterByActive(true)
+		.filterByEffectiveTime(EffectiveTimes.UNSET_EFFECTIVE_TIME)
+		.all()
+		.setExpand("owlRelationships()")
+		.build()
+		.execute(ctx);
+		
+	for (SnomedReferenceSetMember owlAxiomMember : owlAxiomMembers) {
+		def applicableRulePredicates = getApplicableRules(owlAxiomMember.referencedComponent.id)
+			.stream()
+			.map({getPredicate(it)})
+			.filter({ SnomedRelationshipPredicate predicate -> 	
+				def charType = predicate.getCharacteristicTypeId()
+				return Concepts.STATED_RELATIONSHIP == charType || Strings.isNullOrEmpty(charType)
+			}).collect(Collectors.toSet())
+						
+			for (SnomedRelationshipPredicate predicate : applicableRulePredicates) {
+				for (SnomedOWLRelationshipDocument relationship : getOWLRelationships(owlAxiomMember)) {
+					if (relationship.getTypeId().equals(predicate.getAttributeExpression()) && 
+						!getApplicableConcepts(predicate.getRangeExpression()).contains(relationship.getDestinationId())) {
+						issues.add(ComponentIdentifier.of(SnomedTerminologyComponentConstants.REFSET_MEMBER_NUMBER, owlAxiomMember.getId()))
+					}
+				}
+			}
+	}
+	
 } else {
 	for (String domainExpression : mrcmRulesByDomain.keySet()) {
 		Set<String> domain = getApplicableConcepts(domainExpression)
 		
 		for (SnomedConstraint constraint : mrcmRulesByDomain.get(domainExpression)) {
-			SnomedRelationshipPredicate predicate = constraint.getPredicate() instanceof SnomedCardinalityPredicate
-					? ((SnomedCardinalityPredicate) constraint.getPredicate()).getPredicate()
-					: constraint.getPredicate()
+			SnomedRelationshipPredicate predicate = getPredicate(constraint)
 
 			final String attributeExpression = predicate.getAttributeExpression()
 			final String rangeExpression = predicate.getRangeExpression()
@@ -145,7 +184,7 @@ if (params.isUnpublishedOnly) {
 					.filter(SnomedRelationshipIndexEntry.Expressions.active())
 					.filter(SnomedRelationshipIndexEntry.Expressions.sourceIds(domain))
 					.filter(SnomedRelationshipIndexEntry.Expressions.typeId(attributeExpression))
-					.mustNot(SnomedRelationshipIndexEntry.Expressions.destinationIds(getCachedApplicableConcepts(rangeExpression)))
+					.mustNot(SnomedRelationshipIndexEntry.Expressions.destinationIds(getApplicableConcepts(rangeExpression)))
 
 			if (!Strings.isNullOrEmpty(charType)) {
 				expressionBuilder.filter(SnomedRelationshipIndexEntry.Expressions.characteristicTypeId(charType))
@@ -165,6 +204,45 @@ if (params.isUnpublishedOnly) {
 			})
 		}
 	}
+	
+	for (String domainExpression : mrcmRulesByDomain.keySet()) {
+			
+		def applicableMrcmRulePredicates = mrcmRulesByDomain.get(domainExpression).stream()
+			.map({getPredicate(it)})
+			.filter({ SnomedRelationshipPredicate predicate -> 
+				Strings.isNullOrEmpty(predicate.getCharacteristicTypeId()) || Concepts.STATED_RELATIONSHIP.equals(predicate.getCharacteristicTypeId())})
+			.collect(Collectors.toList());		
+		
+		if (applicableMrcmRulePredicates.isEmpty()) {
+			continue
+		}
+		
+		Set<String> domain = getApplicableConcepts(domainExpression)
+		
+		for (SnomedRelationshipPredicate predicate : applicableMrcmRulePredicates) {
+			def attribute = Lists.newArrayList(predicate.getAttributeExpression())
+			def range = getApplicableConcepts(predicate.getRangeExpression())
+			
+			final ExpressionBuilder expressionBuilder = Expressions.builder()
+					.filter(SnomedRefSetMemberIndexEntry.Expressions.active())
+					.filter(SnomedRefSetMemberIndexEntry.Expressions.referencedComponentIds(domain))
+					.filter(SnomedRefSetMemberIndexEntry.Expressions.owlExpressionType(attribute))
+					.mustNot(SnomedRefSetMemberIndexEntry.Expressions.owlExpressionDestination(range))
+			
+			final Query<String> query = Query.select(String.class)
+					.from(SnomedRefSetMemberIndexEntry.class)
+					.fields(SnomedRelationshipIndexEntry.Fields.ID)
+					.where(expressionBuilder.build())
+					.limit(10_000)
+					.build()
+
+			searcher.scroll(query).forEach({ hits ->
+				hits.forEach({ id ->
+					issues.add(ComponentIdentifier.of(SnomedTerminologyComponentConstants.REFSET_MEMBER_NUMBER, id))
+				})
+			})
+		}
+	}
 }
 
-return issues
+return issues as List